requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==36.0.2
Django==3.2.10
django-el-pagination==3.3.0
django-registration-redux==2.10
django-versatileimagefield==2.2
idna==3.6
pillow==10.2.0
psycopg2-binary==2.9.3
pycparser==2.21
python-decouple==3.5
python-magic==0.4.27
pytz==2023.4
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
